[feat] Query Widget made, test widget removed
Removed the poorly implemented test_widget that wrapped an instrument and made its queries and commands accessible.

A more robust and helpful query widget wraps an instruments query.

A command widget should be implemented. Both together (query and command widgets) should be used to create a generic wrapper for instruments.