Incomplete fix for Keithley -420 error
From Prologix manual:
Some instruments generate “Query Unterminated” or “-420” error if they are addressed
to talk after sending a command that does not generate a response (often called non-
query commands). In effect the instrument is saying, I have been asked to talk but I have
nothing to say. The error is often benign and may be ignored. Otherwise, use the
++read command to read the instrument response.

Find a fix for this.